requirements.txt
pillow==10.2.0
ttkthemes==3.2.2
sv-ttk==2.5.5